requirements.txt
Flask==1.1.2
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
gunicorn==20.0.4
itsdangerous==1.1.0
#wsgiref==0.1.2
#flask-wtf==0.14.2
profanity-filter==1.3.3
git+https://github.com/rominf/hunspell_serializable.git
git+https://github.com/kmike/pymorphy2.git
spacy==2.3.5
pymongo==3.11.2
Pillow==9.4.0

opencv-python==4.5.1.48
scipy==1.6.0
numpy~=1.19.2
#tensorflow==2.4.1
#https://git.sr.ht/~johanvandegriff/tensorflow-custom-build/blob/master/tensorflow-2.4.1-cp36-cp36m-linux_x86_64.whl
https://git.sr.ht/~johanvandegriff/tensorflow-custom-build/blob/master/tensorflow-2.4.1-cp38-cp38-linux_x86_64.whl
#tensorflow==2.6.0
